move: MilestoneBlockerCache + ensure_set out of issue/ primitives
Cache parks in local/ (not blocker_interactions as planned: fs_sink (lib)
applies !s at sink time and bin modules are unreachable from it; both land
in tedi_ops at stage 4 anyway). ensure_set inlined at its single call site.
