+++ tedi/src/issue/mod.rs
@@ -10,7 +10,7 @@ mod blocker;
 pub(crate) mod issue_ref;
 pub use issue_ref::IssueRef;
 pub mod milestone_embed;
-pub use blocker::{BlockerItem, BlockerSequence, BlockerSetState, MilestoneBlockerCache, join_with_blockers, split_blockers};
+pub use blocker::{BlockerItem, BlockerSequence, BlockerSetState, join_with_blockers, split_blockers};
 pub use milestone_embed::{MilestoneDoc, parse_blockers_from_embedded, serialize_blockers_view};
 pub use tedi_md::{Events, OwnedCodeBlockKind, OwnedEvent, OwnedTag, OwnedTagEnd};
